json references and tables working now and the route controller is send a valid data to frontend

+++ src/main/java/com/codecool/coolzontations/model/DataFromRequest.java
@@ -12,6 +12,6 @@ import org.springframework.stereotype.Component;
 @Component
 public class DataFromRequest {
 
-    private Integer userID;
-    private Integer consultationID;
+    private Long userID;
+    private Long consultationID;
 }
